Refactor: Move `ProcessResult` and related logic from `pool` to `tracker` package, update references across codebase, and enhance `ProgressTracker` with results handling.
