+++ internal/tracker/result.go
@@ -1,4 +1,4 @@
-package pool
+package tracker
 
 type ProcessStatus int
