+++ internal/tracker/result_test.go
@@ -1,4 +1,4 @@
-package pool
+package tracker
 
 import (
 	"errors"
